requirements.txt
pandas>=2.0.0
python-slugify>=8.0.0
python-frontmatter>=1.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
PyYAML>=6.0.0
